Added OnServiceUpdate to HMI subscriptions
-Moved notification parameters to msg_params
-Added service_event to notification on StartStreaming
-Added OnServiceUpdate to preloaded_pt.json

+++ src/components/application_manager/src/message_helper/message_helper.cc
@@ -1998,8 +1998,7 @@ MessageHelper::CreateOnServiceStatusUpdateNotification(
 
   auto& msg_ref = (*on_status_update_notification);
 
-  msg_ref[strings::msg_params][hmi_notification::service_type] =
-      (uint32_t)service_type;
+  msg_ref[strings::msg_params][hmi_notification::service_type] = service_type;
 
   return on_status_update_notification;
 }
